Moved the skill and skilltype entities to the hydra-client entities folder.
