--- client-side/src/app/hydra-client/entities/HydraSkill.ts
@@ -1,5 +1,0 @@
-export class HydraSkill {
-  skillId: number;
-  skill: string;
-  active: boolean;
-}
