+++ client-side/src/app/hydra-client/entities/Skill.ts
