+++ client-side/src/app/portals/Caliber/assess/assess.component.ts
@@ -22,7 +22,7 @@ import { HydraBatch } from '../../../hydra-client/entities/HydraBatch';
 import { HydraBatchUtilService } from '../../../services/hydra-batch-util.service';
 import { HydraTrainee } from '../../../hydra-client/entities/HydraTrainee';
 import { SkillService } from '../services/skill.service';
-import { Skill } from '../../../entities/Skill';
+import { Skill } from '../../../hydra-client/entities/Skill';
 
 @Component({
   selector: 'app-assess',
